// src/types/user.type.ts
// user.type.ts
// -----------------------------------------------------------

export interface IUser  {
    _id: string,
    usr_name: string;
    usr_email: string;
    usr_street_address: string;
    state_id: string;
    city_id: string;
    usr_zip: string;
    usr_phone_number: string;
    usr_username: string;
    usr_password: string;
    rol_id: string;
  }

// -----------------------------------------------------------